+++ src/components/atoms/Subtitle/styles/index.ts
@@ -1,1 +1,1 @@
-export * from './SubTitle.style';
+export * from './Subtitle.style';
